refactor: move v2 inspection tests to correct package
