+++ src/test/kotlin/com/asyncapi/plugin/idea/inspection/v2/_3_0/AsyncAPIJsonSpecificationInspectionTest.kt
@@ -1,13 +1,13 @@
-package com.asyncapi.plugin.idea.inspection.v2_6_0
+package com.asyncapi.plugin.idea.inspection.v2._3_0
 
 import com.asyncapi.plugin.idea.inspection.AbstractAsyncAPIJsonSpecificationInspectionTest
 
 /**
  * @author Pavel Bodiachevskii
- * @since 2.3.0
+ * @since 1.6.0
  */
 class AsyncAPIJsonSpecificationInspectionTest: AbstractAsyncAPIJsonSpecificationInspectionTest() {
 
-    override fun getTestDataPath(): String = "src/test/testData/json/inspection/2.6.0"
+    override fun getTestDataPath(): String = "src/test/testData/json/inspection/2.3.0"
 
 }
