+++ src/test/testData/json/inspection/2.99.99/without channels.json
